refactor order-checker tests and add optional parametr in two methods
