+++ package.json
@@ -37,6 +37,7 @@
     "crypto-js": "^3.1.9-1",
     "express": "^4.15.2",
     "mongoose": "^4.8.5",
+    "nock": "^9.0.13",
     "precommit-hook": "^3.0.0",
     "query-string": "^4.3.2",
     "standard": "^9.0.1",
